// front/src/app/shared/interfaces/entities.interface.ts
export enum PostStatuses {
  Draft = 'draft',
  Pub = 'pub',
  Del = 'del',
  Private = 'private',
}
export enum CategoryStatuses {
  Pub = 'pub',
  Private = 'private',
}
export enum UserRoles {
  OWNER = 'owner',
  CREATOR = 'creator',
  VISITOR = 'visitor',
}
export enum UserStatuses {
  ACTIVE = 'active',
  BANNED = 'banned',
  DELETED = 'deleted',
}